added lovalStorage, tags and helpers.js
